Updated stop codon detection
Cases where the reference sequence and the sequence to be annotated hasn't been handled correctly for a LOOONNNGG time. Fixed at least a few of the edge cases with this update. Now I allow proteins to be up to 10 amino acids longer than the reference and any amount shorter (although this in almost all cases indicates a bad consensus sequence). Going to more seriously debug this next week - but basic functionality should be much improved.