v1.6.4
Fixed edge case where proteins near the end of the genome that needed to be shifted to the left would break. (when the submitted sequence is shorter than the reference sequence and the last protein is close to the end of the genome). This update also implemented the weird false typo from commit [7e41b87](https://github.com/rcs333/VAPiD/commit/7e41b871b89b815969e463919957b16fc0f5cb68)